requirements.txt
Django==5.1.2
djangorestframework==3.14.0
gunicorn==20.1.0
django-cors-headers==3.13.0
crewai
langchain-groq
crewai-tools
python-dotenv
